Concept map equivalence types

The following values are supported:

RELATEDTO - The concepts are related to each other, and have at least some overlap in meaning, but the exact relationship is not known.

EQUIVALENT - The definitions of the concepts mean the same thing (including when structural implications of meaning are considered) (i.e. extensionally identical).

EQUAL - The definitions of the concepts are exactly the same (i.e. only grammatical differences) and structural implications of meaning are identical or irrelevant (i.e. intentionally identical).

WIDER - The target mapping is wider in meaning than the source concept.

SUBSUMES - The target mapping subsumes the meaning of the source concept (e.g. the source is-a target).

NARROWER - The target mapping is narrower in meaning than the source concept. The sense in which the mapping is narrower SHALL be described in the comments in this case, and applications should be careful when attempting to use these mappings operationally.

SPECIALIZES - The target mapping specializes the meaning of the source concept (e.g. the target is-a source).

INEXACT - There is some similarity between the concepts, but the exact relationship is not known.

UNMATCHED - This is an explicit assertion that there is no mapping between the source and target concept.

DISJOINT - This is an explicit assertion that the target concept is not in any way related to the source concept.

Equivalence

Format

An object of class list of length 10.

See also