Functions to perists data sources in various formats.

ds_write_ndjson() writes the data to a directory of NDJSON files. The files will be named using the resource type and the ".ndjson" extension.

ds_write_parquet() writes the data to a directory of Parquet files.

ds_write_delta() writes the data to a directory of Delta files.

ds_write_tables() writes the data to a set of tables in the Spark catalog.

ds_write_ndjson(ds, path, file_name_mapper = NULL)

ds_write_parquet(ds, path)

ds_write_delta(ds, path, import_mode = ImportMode$OVERWRITE)

ds_write_tables(ds, schema = NULL, import_mode = ImportMode$OVERWRITE)

Arguments

ds

The DataSource object.

path

The URI of the directory to write the files to.

file_name_mapper

An optional function that can be used to customise the mapping of the resource type to the file name. Currently not implemented.

import_mode

The import mode to use when writing the data - "overwrite" will overwrite any existing data, "merge" will merge the new data with the existing data based on resource ID.

schema

The name of the schema to write the tables to.

Examples

if (FALSE) { # pathling_is_spark_installed()
# Create a temporary warehouse location, which will be used when we call ds_write_tables().
temp_dir_path <- tempfile()
dir.create(temp_dir_path)
sc <- sparklyr::spark_connect(master = "local[*]", config = list(
  "sparklyr.shell.conf" = c(
    paste0("spark.sql.warehouse.dir=", temp_dir_path),
    "spark.sql.extensions=io.delta.sql.DeltaSparkSessionExtension",
    "spark.sql.catalog.spark_catalog=org.apache.spark.sql.delta.catalog.DeltaCatalog"
  )
), version = pathling_spark_info()$spark_version)

pc <- pathling_connect(sc)
data_source <- pc %>% pathling_read_ndjson(pathling_examples('ndjson'))

# Write the data to a directory of NDJSON files.
data_source %>% ds_write_ndjson(file.path(tempdir(), 'ndjson'))

# Write the data to a directory of Parquet files.
data_source %>% ds_write_parquet(file.path(tempdir(), 'parquet'))

# Write the data to a directory of Delta files.
data_source %>% ds_write_delta(file.path(tempdir(), 'delta'), import_mode = ImportMode$OVERWRITE)

# Write the data to a set of Spark tables in 'fhir' database.
data_source %>% ds_write_tables("default", import_mode = ImportMode$MERGE)

pathling_disconnect(pc)
unlink(temp_dir_path, recursive = TRUE)
}