Write FHIR data to Delta files

Writes the data from a data source to a directory of Delta files.

ds_write_delta(ds, path, import_mode = ImportMode$OVERWRITE)

Arguments

ds

The DataSource object.

path

The URI of the directory to write the files to.

import_mode

The import mode to use when writing the data - "overwrite" will overwrite any existing data, "merge" will merge the new data with the existing data based on resource ID.

Value

No return value, called for side effects only.

See also

Examples

pc <- pathling_connect()
data_source <- pc %>% pathling_read_ndjson(pathling_examples('ndjson'))

# Write the data to a directory of Delta files.
data_source %>% ds_write_delta(file.path(tempdir(), 'delta'), import_mode = ImportMode$OVERWRITE)

pathling_disconnect(pc)