Test subsumption between codings

Takes two Coding columns as input. Returns a Column that contains a Boolean value, indicating whether the left Coding subsumes the right Coding.

tx_subsumes(left_codings, right_codings)

Arguments

left_codings

A Column containing a struct representation of a Coding or an array of Codings.

right_codings

A Column containing a struct representation of a Coding or an array of Codings.

Value

A Column containing the result of the operation (boolean).

See also

Examples

pc <- pathling_connect()

# Test the codings of the Condition `code` for subsumption of a SNOMED CT code.
pc %>% pathling_example_resource('Condition') %>%
    sparklyr::mutate(
         id,
         subsumes = !!tx_subsumes(code[['coding']],
             !!tx_to_snomed_coding('444814009')),
         .keep='none')
#> Error: org.apache.spark.sql.AnalysisException:
#> [UNRESOLVED_COLUMN.WITH_SUGGESTION] A column or function parameter with name
#> `Condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`*` cannot be resolved. Did
#> you mean one of the following?
#> [`condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`id`,
#> `condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`_fid`,
#> `condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`code`,
#> `condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`meta`,
#> `condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`.`note`].; line 2 pos 2;
#> 'Project ['Condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de.*,
#> subsumes(code#25.coding, CASE WHEN NOT isnull(444814009) THEN struct(col1,
#> null, col2, cast(http://snomed.info/sct as string), col3, cast(null as string),
#> col4, cast(444814009 as string), col5, null, col6, null) END, false) AS
#> subsumes#2186] +- SubqueryAlias condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de
#> +- View (`Condition_c7dab40d_3c57_41f7_9edb_bf9036a2b1de`,
#> [id#14,id_versioned#15,meta#16,implicitRules#17,language#18,text#19,identifier#20,clinicalStatus#21,verificationStatus#22,category#23,severity#24,code#25,bodySite#26,subject#27,encounter#28,onsetAge#29,onsetDateTime#30,onsetPeriod#31,onsetRange#32,onsetString#33,abatementAge#34,abatementDateTime#35,abatementPeriod#36,abatementRange#37,abatementString#38,recordedDate#39,recorder#40,asserter#41,stage#42,evidence#43,note#44,_fid#45,_extension#46])
#> +- Relation
#> [id#14,id_versioned#15,meta#16,implicitRules#17,language#18,text#19,identifier#20,clinicalStatus#21,verificationStatus#22,category#23,severity#24,code#25,bodySite#26,subject#27,encounter#28,onsetAge#29,onsetDateTime#30,onsetPeriod#31,onsetRange#32,onsetString#33,abatementAge#34,abatementDateTime#35,abatementPeriod#36,abatementRange#37,...
#> 9 more fields] parquet
#> 
#> Run `sparklyr::spark_last_error()` to see the full Spark error (multiple lines)
#> To use the previous style of error message set
#> `options("sparklyr.simple.errors" = TRUE)`
 
pathling_disconnect(pc)